requirements.txt
streamlit==1.32.0
pandas==2.2.1
google-generativeai>=0.7.2
openpyxl==3.1.2
matplotlib==3.8.3
scipy==1.12.0
statsmodels==0.14.1
streamlit-paste-button
